- KGamePropertyList: kdDebug() shouldn't give the used data. the debug "<<"   isn't always implemented - KPlayer/COMPAT: KPlayer::signalNetworkData MUST use const QByteArray&, NOT   QDataStream& - adding Martin as @author of KGameIOs
svn path=/trunk/kdegames/libkdegames/; revision=118383

+++ kgame/kgame.cpp
@@ -981,7 +981,7 @@ void KGame::networkTransmission(QDataStream &stream, int msgid, Q_UINT32 receive
                       << endl;
      }
      kdDebug(11001) << "KGame::slotNetworkTransmision:: User data msgid " << msgid << endl;
-     emit signalNetworkData(msgid - KGameMessage::IdUser,((QBuffer*)stream.device())->readAll(),(int)receiver,sender);
+     emit signalNetworkData(msgid - KGameMessage::IdUser,((QBuffer*)stream.device())->readAll(),receiver,sender);
    }
    break;
  }
